requirements.txt

click==6.7
requests>=2.20.0
Django==2.2.7
djangorestframework==3.9.2
django-extensions==2.2.5
